[core] rename program to course
