+++ packages/core/src/models/course/category.schema.ts
