+++ packages/core/src/models/course/category.ts
