+++ packages/core/src/models/course/competency.schema.ts
