+++ packages/core/src/models/course/competency.test.ts
