+++ packages/core/src/models/course/competency.ts
