+++ packages/core/src/models/course/degree.schema.ts
