+++ packages/core/src/models/course/degree.test.ts
