+++ packages/core/src/models/course/degree.ts
