+++ packages/core/src/models/course/discipline.schema.ts
