+++ packages/core/src/models/course/discipline.test.ts
