+++ packages/core/src/models/course/discipline.ts
