+++ packages/core/src/models/course/index.ts
@@ -3,4 +3,4 @@ export * as Competency from './competency.js'
 export * as Degree from './degree.js'
 export * as Discipline from './discipline.js'
 export * as Module from './module.js'
-export * from './program.js'
+export * as Program from './program.js'
