+++ packages/core/src/models/course/module.schema.ts
