+++ packages/core/src/models/course/module.ts
