+++ packages/core/src/models/course/program.schema.ts
