+++ packages/core/src/models/course/program.test.ts
