+++ packages/core/src/models/course/program.ts
