+++ packages/core/src/models/curriculum/curriculum.test.ts
@@ -2,8 +2,8 @@ import assert from 'assert'
 import test from 'node:test'
 import { withTestTransaction } from '../../contexts/index.js'
 import dayjs from '../../utils/dayjs.js'
+import { Degree, Discipline } from '../course/index.js'
 import { Program } from '../index.js'
-import { Degree, Discipline } from '../program/index.js'
 import * as Curriculum from './curriculum.js'
 
 test.skip('curriculum crud', () =>
