+++ packages/core/src/models/index.ts
@@ -1,8 +1,8 @@
 export * as ApiKey from './api-key/index.js'
 export * as Certificate from './certificate/index.js'
+export * as Program from './course/index.js'
 export * as Curriculum from './curriculum/index.js'
 export * as Location from './location/index.js'
 export * as Platform from './platform/index.js'
-export * as Program from './program/index.js'
 export * as Student from './student/index.js'
 export * as User from './user/index.js'
